Using the pwiz code with Boost versions other than the one we provide turns out to be a bit ugly.  Boost filesystem v2 support is going away, and v3 breaks the API in surprising ways - see http://www.boost.org/doc/libs/1_47_0/libs/filesystem/v3/doc/deprecated.html In BFS2 p.filename() or p.leaf() or p.extension() returns a string, but in BFS3 they return a bfs::path object. In BFS2 complete() is in bfs namespace, but in BFS3 it's not in a namespace at all. So, a handful of ifdefs and preprocessor hijinks are needed.
Former-commit-id: cd9343f982270941a01ff6a2bc672f6451695d69

+++ pwiz/data/msdata/BinaryDataEncoderTest.cpp
@@ -233,7 +233,7 @@ void testBadFile(const string& filename)
 
     try
     {
-        filesize = bfs::file_size(filename);
+        filesize = (size_t) bfs::file_size(filename);
     }
     catch (exception&)
     {
